feat: Add CRUD functionality for Sampah Terkelola and Tujuan Sampah
- Implemented edit, add, and index views for Sampah Terkelola with form validation and confirmation prompts.
- Created index view for displaying Sampah Terkelola data in a table with delete confirmation.
- Added edit, add, and index views for Tujuan Sampah with similar structure and validation.
- Developed user management views including edit, add, and index with role selection and validation.
- Enhanced UI with consistent styling and responsive design.

+++ database/migrations/2025_10_07_013921_create_sampah_terkelolas_table.php
@@ -16,7 +16,7 @@ return new class extends Migration
             $table->foreignId('id_user')->constrained('users');
             $table->foreignId('id_lokasi')->constrained('lokasi_asals');
             $table->foreignId('id_jenis')->constrained('jenis');
-            $table->integer('jumlah_berat');
+            $table->decimal('jumlah_berat', 10, 2);
             $table->date('tgl');
             $table->string('foto')->nullable();
             $table->timestamps();
